Move rap archive to resque, first step on moving all steps to resque
For now the differences are that the archive worker will run inside
resque (it is a resque worker now) and there is a "rap-trigger" file that
is executed by systemd to detect recordings that need to be archived and
add a job on resque to archive them.

The process os scheduling jobs still needs to be reviewed, but for now will
be done by rap-trigger.